Feat: 전시 상세페이지 토글 스위치 버튼 공통 컴포넌트로 따로 분리
